Fix test_prio_inherit.sh hanging indefinitely
The pthread_test program runs in an infinite loop until it receives
SIGINT. The test script was calling it directly without sending a
signal, causing it to hang forever.

Changed to run pthread_test in background, let it run for 5 seconds,
then send SIGINT and wait for clean exit.

+++ tests/priority-inheritance/test_prio_inherit.sh
@@ -17,9 +17,11 @@ EVENTS="sched_wakeup* sched_switch sched_migrate* sched_pi_setprio"
 CPUSET_DIR=/sys/fs/cgroup
 
 tear_down() {
-  trace_write "kill $PID"
-  kill -9 $PID
-  
+  if [ -n "$PID" ] && kill -0 $PID 2>/dev/null; then
+    trace_write "kill $PID"
+    kill -9 $PID 2>/dev/null
+  fi
+
   trace_stop
   trace_extract
 }
@@ -32,9 +34,25 @@ trace_start
 trace_write "TEST $TNAME START"
 
 trace_write "Launch pthread_test [inherit]"
-./pthread_test inherit
+./pthread_test inherit &
+PID=$!
 
-trace_write "PASS"
+trace_write "pid: $PID"
+trace_write "Sleep for 5s to let test run"
+sleep 5
+
+trace_write "Sending SIGINT to pthread_test"
+kill -INT $PID
+wait $PID
+RES=$?
+
+if [ $RES -eq 0 ]; then
+  trace_write "PASS"
+else
+  trace_write "FAIL: pthread_test exited with $RES"
+  tear_down
+  exit 1
+fi
 trace_write "TEST $TNAME FINISH"
 tear_down
 sleep 1
